third_party/blink/web_tests/platform/generic/fast/js/mozilla/strict/15.4.4.13-expected.txt

PASS 'use strict'; var a = arr(); [a.unshift(40, 50), a] threw exception of type TypeError.
PASS var a = arr(); [a.unshift(40, 50), a] threw exception of type TypeError.
PASS true === true
PASS 'use strict'; var o = obj(); [Array.prototype.unshift.call(o, 40, 50), o] threw exception of type TypeError.
PASS var o = obj(); [Array.prototype.unshift.call(o, 40, 50), o] threw exception of type TypeError.
PASS true === true
PASS 'use strict'; var a = agap(); [a.unshift(9), a] threw exception of type TypeError.
PASS var a = agap(); [a.unshift(9), a] threw exception of type TypeError.
PASS true === true
PASS 'use strict'; var o = ogap(); [Array.prototype.unshift.call(o, 9), o] threw exception of type TypeError.
PASS var o = ogap(); [Array.prototype.unshift.call(o, 9), o] threw exception of type TypeError.
PASS true === true
 PASSED! 
PASS successfullyParsed is true

TEST COMPLETE
